revert changes made to add fixture to new conftest file
